+++ tests/x509/test_x509_ext.py
@@ -3271,13 +3271,6 @@ class TestAuthorityKeyIdentifierExtension(object):
         ski_ext = issuer_cert.extensions.get_extension_for_class(
             x509.SubjectKeyIdentifier
         )
-        # This was the incorrect arg we want to deprecate and remove
-        with pytest.warns(utils.CryptographyDeprecationWarning):
-            aki = x509.AuthorityKeyIdentifier.\
-                from_issuer_subject_key_identifier(ski_ext)
-            assert ext.value == aki
-
-        # Here's what we actually documented and want to do
         aki = x509.AuthorityKeyIdentifier.from_issuer_subject_key_identifier(
             ski_ext.value
         )